Major refactoring to the SurveyBuilder/SurveyElement JS.
SurveyElement is now not aware of any other part of the DOM except
itself. SurveyBuilder is a collection of SurveyElements and
orchestrates interactions between them.

Added Jasmine & Capybara and specs for these scenarios.

This is a significant change to the structure of the JS.
Hope you don't have any merge conflicts. :)