Finish styling the Surveys index page.
This is a combination of 4 commits.

The first commit's message is:
Change the font for the `published/unpublished` links.

This is the 2nd commit message:
Show a message in the index page when there are no surveys to display.

This is the 3rd commit message:
Hide the `published/unpublished` filter links when there are no surveys
to display.

This is the 4th commit message:
Add french translation for the 'Create Survey' link