Responses for a public survey can be added by a user who isn't logged in.
Since the controller isn't strictly RESTful, and there isn't a `new` action, so the authorization for the `edit` action had to be removed for responses to public surveys.
This means that a non-logged-in user can edit ANY response for a public survey by editing the number in the URL. Not sure how to remedy this without rewriting the controller in a RESTful way.