Use the fat arrow (=>) for all methods.
A lot of processing is happening on the UI thread, so it'll be good to
do a `_.defer` so a lot of these happen on a separate thread.

If we do a defer, having the fat arrow everywhere is important so that
`this` is bound correctly.

Revert this if we decide not to use `_.defer`.